Add key handling to 3DEN attribute
